Use the home_assistant user for most tasks
This updates the formula to use the home_assistant user for tasks like
cloning the git repo and making the virtual environment. This allows
home assistant to perform some tasks it otherwise can't do due to
permission issues, like installing needed python libraries at runtime.

This also allows the ssh identity used for cloning the git repo to be
specified.

Signed-off-by: M. David Bennett <mdavidbennett@syntheticworks.com>